% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_history.R
\name{bp_is_fraction_tick}
\alias{bp_is_fraction_tick}
\title{Test Current Tick Fraction}
\usage{
bp_is_fraction_tick(state, fraction, tol = 1e-08)
}
\arguments{
\item{state}{Program state.}

\item{fraction}{Numeric fraction within the year, e.g. `0`, `0.25`, `0.5`.}

\item{tol}{Numeric tolerance.}
}
\value{
Logical scalar.
}
\description{
Test Current Tick Fraction
}
