% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synthetic_traits.R
\name{bp_register_synthetic_traits}
\alias{bp_register_synthetic_traits}
\title{Register Synthetic Traits}
\usage{
bp_register_synthetic_traits(state, synthetic_traits)
}
\arguments{
\item{state}{Program state.}

\item{synthetic_traits}{One definition or a list of definitions.}
}
\value{
Updated program state.
}
\description{
Register Synthetic Traits
}
