% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_history.R
\name{bp_reset_costs}
\alias{bp_reset_costs}
\title{Reset Cost Log}
\usage{
bp_reset_costs(state)
}
\arguments{
\item{state}{Program state.}
}
\value{
Updated program state.
}
\description{
Reset Cost Log
}
