% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_history.R
\name{bp_tick_fraction}
\alias{bp_tick_fraction}
\title{Current Tick Fraction Within Year}
\usage{
bp_tick_fraction(state)
}
\arguments{
\item{state}{Program state.}
}
\value{
Numeric fraction in `[0, 1)`.
}
\description{
Current Tick Fraction Within Year
}
