% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_history.R
\name{bp_ticks_per_year}
\alias{bp_ticks_per_year}
\title{Ticks per Year}
\usage{
bp_ticks_per_year(state)
}
\arguments{
\item{state}{Program state.}
}
\value{
Integer number of simulation ticks per calendar year.
}
\description{
Ticks per Year
}
