% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_history.R
\name{bp_update_stage_pop}
\alias{bp_update_stage_pop}
\title{Update an Existing Stage Population}
\usage{
bp_update_stage_pop(
  state,
  cohort_id,
  pop,
  require_same_ids = TRUE,
  allow_reorder = FALSE,
  log_event = TRUE
)
}
\arguments{
\item{state}{Program state.}

\item{cohort_id}{Cohort id to update.}

\item{pop}{Replacement population object.}

\item{require_same_ids}{Require same individual ids.}

\item{allow_reorder}{Allow replacement to be reordered to the original order.}

\item{log_event}{Whether to append an update event.}
}
\value{
Updated program state.
}
\description{
Safely replace a stored population object after validating population size and
individual identity/order.
}
