% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/state_history.R
\name{bp_years_to_ticks}
\alias{bp_years_to_ticks}
\title{Convert Years to Ticks}
\usage{
bp_years_to_ticks(state, years, require_integer = TRUE, tol = 1e-08)
}
\arguments{
\item{state}{Program state.}

\item{years}{Single non-negative numeric duration in years.}

\item{require_integer}{Whether the duration must map to an integer number of
ticks.}

\item{tol}{Tolerance used when checking integer representability.}
}
\value{
Integer number of ticks.
}
\description{
Convert a calendar duration in years to simulation ticks using
`state$time$dt`.
}
