% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readable_api.R
\name{pop_n_ind}
\alias{pop_n_ind}
\title{Count Individuals in a Population-Like Object}
\usage{
pop_n_ind(pop)
}
\arguments{
\item{pop}{Population-like object.}
}
\value{
Integer number of individuals.
}
\description{
Return the number of individuals represented by an AlphaSimR population,
matrix, data frame, or another object that implements [nrow()].
}
